+++ lib/business_logic.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "business_logic/version"
+require_relative "business_logic/abstract_method_error"
 require_relative "business_logic/railtie" if defined?(Rails::Railtie)
 require_relative "business_logic/form"
 require_relative "business_logic/seed"